# backend/docs/openapi.yaml
openapi: 3.0.3
...
info:
  title: WIT API
  version: "1.0.0"
  description: |
    Backend for the WIT deck catalog — a Netflix-style library of presentation
    decks. Go + PostgreSQL, JSON over HTTP.

    ## Authentication

    `POST /auth/login` returns a JWT (HS256, 24h by default — `JWT_TTL`). Send it
    as `Authorization: Bearer <token>`.

    Three roles, enforced at the router:

    | Role     | Can do                                    |
    |----------|-------------------------------------------|
    | `admin`  | everything, including managing users      |
    | `editor` | create/update/delete decks, upload files  |
    | `viewer` | read, and manage their own favourites     |

    Reads on `/decks` and `/users` are public. A 401 on an authenticated call
    means the token is missing, malformed, or expired.

    ## Errors

    Every failure uses the same envelope, so a client can branch on `code`
    without parsing prose:

    ```json
    { "error": { "code": "not_found", "message": "deck not found" } }
    ```

    Codes: `invalid_input` (400), `unauthorized` (401), `email_not_verified`
    (401), `forbidden` (403), `not_found` (404), `conflict` (409),
    `rate_limited` (429), `internal` (500).

    ## Rate limiting

    `/auth/*` is limited on two axes at once — per calling address, and per
    account named in the body. Either alone leaves a hole: per-IP misses a
    botnet against one admin, per-account misses one host spraying a password
...

...

    ## Paging

...

...

...
servers:
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: admin@wit.id }
                password: { type: string, format: password, example: admin1234 }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: JWT, example: "eyJhbGciOiJIUzI1NiIs..." }
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401":
          description: |
            Wrong password, unknown email, or a suspended account — deliberately
            indistinguishable, since telling them apart would confirm which
            emails have accounts.

            A correct password on an unverified account returns `401` with code
            `email_not_verified` instead. That is checked *after* the password,
            so it never reveals whether an address is registered; clients turn
            it into "check your inbox" with a resend button.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      security: []
      description: |
        Creates an unverified `viewer` and emails a verification link.

        **Role and status are not read from the body.** An unauthenticated caller
        choosing their own role would be a straight privilege escalation, so a
        `"role": "admin"` in the request is silently ignored and the account is
        always created as `viewer`.

        No token is returned: the account cannot sign in until the address is
        verified, so issuing one here would defeat the point. A duplicate email
        is a plain `409` — address enumeration is not defensible while
        `GET /users` is public and lists every address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:     { type: string, example: Ada Lovelace }
                email:    { type: string, format: email, example: ada@wit.id }
                password: { type: string, format: password, minLength: 8 }
      responses:
        "201":
          description: |
            Account created, verification email sent. Also returned when the
            account was created but the email failed to send — the account
            already exists at that point, and failing the request would strand
            an address nobody can verify. The client offers "resend".
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:    { $ref: "#/components/schemas/User" }
                  message: { type: string }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "409":
          description: Email already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/verify:
    post:
      tags: [Auth]
      summary: Redeem a verification token
      security: []
      description: |
        Marks the address proven and signs the user straight in — they have just
        demonstrated control of the address and chose a password minutes ago, so
        asking them to retype it adds friction without adding a check.

        Tokens are single-use and expire after 24 hours. Verifying also burns
        every other outstanding token for that user, so a link from an earlier
        resend cannot be replayed later.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: The value from the `token` query parameter of the emailed link }
      responses:
        "200":
          description: Verified and signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: JWT }
                  user:  { $ref: "#/components/schemas/User" }
        "400":
          description: |
            Missing, unknown, already-used, or expired token. Expiry is reported
            distinctly from invalid — an expired link means the account is real
            and a resend will work, which is worth telling someone.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Send a fresh verification link
      security: []
      description: |
        Always returns `204`, whether or not the address is registered and
        whether or not it is already verified. Unlike register this takes only
        an address, so a truthful "no such account" would turn it into an oracle
        for any address — and unlike register there is no usability cost,
        because someone who just signed up knows they did.

        Limited to 5 verification emails per account per hour. Each one is
        outbound mail to an address that may not belong to the requester.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204": { description: Handled — no information about the address is returned }
        "400":
          description: Malformed email, or the hourly resend limit was reached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /decks:
    get:
      tags: [Decks]
      summary: List decks (paged)
      security: []
      description: |
        Filters combine with AND. `search` matches title, subtitle, author and
        description, case-insensitively.
      parameters:
        - { name: search,     in: query, schema: { type: string }, description: "Substring match across title/subtitle/author/description" }
        - { name: category,   in: query, schema: { type: string }, example: engineering }
        - { name: industry,   in: query, schema: { type: string }, example: tech }
        - { name: sourceType, in: query, schema: { type: string, enum: [pdf, url, video, gslides, embed] } }
        - { name: featured,   in: query, schema: { type: boolean } }
        - name: ids
          in: query
          schema: { type: string }
          description: |
            Comma-separated deck ids, for hydrating a known set (favourites,
            continue-watching) without listing the catalog. Capped at 200. An
            `ids` parameter that resolves to nothing returns `[]` rather than
            falling back to a full listing.
          example: "0f7c...,3b21..."
        - name: sort
          in: query
          schema: { type: string, enum: [newest, oldest, views, title], default: newest }
          description: |
            Every ordering is tie-broken by id. Without that, rows sharing a sort
            key can reorder between pages and a paging client silently skips or
            repeats them.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          description: Values above 200 are clamped, not rejected.
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: A page of decks
          headers:
            X-Total-Count: { schema: { type: integer }, description: Total matches, ignoring paging }
            X-Limit:       { schema: { type: integer }, description: Page size actually applied }
            X-Offset:      { schema: { type: integer } }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/InvalidInput" }
    post:
      tags: [Decks]
      summary: Create a deck
      description: Requires `admin` or `editor`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeckInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /decks/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Decks]
      summary: Fetch one deck
      security: []
      responses:
        "200":
          description: The deck
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Decks]
      summary: Update a deck (partial)
      description: |
        Requires `admin` or `editor`. Every field is optional — omitted fields
        are left untouched, so two clients editing different fields don't
        clobber each other. Sending `source` replaces it wholesale.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeckPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Decks]
      summary: Delete a deck
      description: |
        Requires `admin` or `editor`. Favourite rows referencing the deck are
        removed with it (`ON DELETE CASCADE`).
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /decks/{id}/views:
    post:
      tags: [Decks]
      summary: Increment the view counter
      security: []
      description: |
        Public and unauthenticated — view counts are a global signal, not
        per-user history. Fire-and-forget: clients should not block playback on
        the response.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The deck, with the new count
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "404": { $ref: "#/components/responses/NotFound" }
  /audit:
    get:
      tags: [Activity]
      summary: Who changed what
      description: |
        Requires `admin`. The log names every account and what they changed,
        which is the picture an attacker wants and nobody else needs.

        Every write is recorded, and only writes — reads would bury them and
        the interesting question is what changed. Refused requests are left
        out: an attempt that changed nothing is not a change, and recording
        every one of them makes the real ones hard to find.

        The actor's email is copied into each entry when it is written rather
        than joined on read, so deleting an account does not erase its own
        history.

        Paged like `/decks`: the body is a plain array and the total rides in
        `X-Total-Count`.
      parameters:
        - name: entity
          in: query
          schema: { type: string, enum: [deck, user, taxonomy, settings, upload, profile, auth, favorites, progress] }
        - name: action
          in: query
          schema: { type: string, enum: [create, update, delete] }
        - name: actorId
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: One page of entries, newest first
          headers:
            X-Total-Count:
              schema: { type: integer }
              description: Entries matching the filter, ignoring paging
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:         { type: integer, format: int64 }
                    at:         { type: string, format: date-time }
                    actorId:    { type: string, format: uuid, nullable: true }
                    actorEmail: { type: string, description: Copied when written, so it survives the account }
                    actorRole:  { type: string }
                    action:     { type: string, enum: [create, update, delete] }
                    entity:     { type: string }
                    entityId:   { type: string }
                    route:      { type: string, description: "The matched route, e.g. PUT /decks/{id}" }
                    status:     { type: integer }
                    ip:         { type: string }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /me:
    get:
      tags: [Profile]
      summary: The signed-in account
      description: |
        Any signed-in role. Separate from `/users`, which is admin only and
        returns everybody: a viewer has every right to read their own record
        and no business reading anyone else's.

        `deckCount` counts decks created by this account. Decks added before
        `created_by` existed are attributed to nobody, so an established
        catalog can show a count of zero for someone who did in fact add to it.
      responses:
        "200":
          description: The account, with the numbers a profile page shows
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/User"
                  - type: object
                    properties:
                      deckCount: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me/password:
    put:
      tags: [Profile]
      summary: Change your own password
      description: |
        Requires the current password. Deliberately not the same path as an
        admin resetting somebody else's through `PUT /users/{id}`: an admin has
        no current password to offer, and a person changing their own must not
        be able to skip proving they know it. Without that step a stolen token
        could replace the password outright, which is the difference between
        losing a session and losing an account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword:     { type: string, format: password, minLength: 8 }
      responses:
        "204": { description: Changed }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401":
          description: Missing token, or the current password is wrong
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /settings:
    get:
      tags: [Settings]
      summary: Read the application settings
      description: |
        Public. The navigation reads these before anyone has signed in, and
        none of the values are secret.

        Always complete: every known key is returned with its default filled
        in, so a client never has to know which keys exist in order to read
        them.
      responses:
        "200":
          description: Every setting, defaults applied
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
              example:
                nav_max_categories: "5"
    put:
      tags: [Settings]
      summary: Change settings
      description: |
        Requires `admin`. Partial: keys left out keep their current value.

        Unknown keys are refused rather than stored. A key nothing reads is a
        setting somebody believes they changed, and the failure would be
        silent. Values are validated against what their key means, and the
        whole body is checked before any of it is written — a half-applied
        batch leaves the settings in a state nobody asked for.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { type: string }
              properties:
                nav_max_categories:
                  type: string
                  description: |
                    How many categories the header shows before the rest move
                    into its "More" menu. A whole number from 1 to 20; the
                    frontend still drops any that do not fit the window.
      responses:
        "200":
          description: The settings after the change
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /taxonomy/{kind}:
    parameters:
      - $ref: "#/components/parameters/TaxonomyKind"
    get:
      tags: [Taxonomy]
      summary: List the terms of one master list
      description: |
        Public. These are the lists the catalog is browsed by; they were
        compiled into the frontend bundle until migration `000010` moved them
        here, so changing one meant a deploy.

        Retired terms are included by default. Pass `active=true` for the
        browse UI; the admin screen needs the full list, because a term you
        cannot see is a term you cannot bring back.
      parameters:
        - name: active
          in: query
          schema: { type: string, enum: ["true"] }
          description: Only terms still offered for new decks.
      responses:
        "200":
          description: Terms, ordered by `sortOrder` then `slug`
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TaxonomyTerm" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Taxonomy]
      summary: Add a term
      description: |
        Requires `admin`. Editors create decks and pick from these lists;
        changing the lists changes the site's own navigation.

        Omitting `sortOrder` places the term at the end rather than at zero,
        which would silently reorder the home page.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, title]
              properties:
                slug:
                  type: string
                  pattern: "^[a-z0-9][a-z0-9-]{0,63}$"
                  description: |
                    What decks store. Chosen once — it cannot be edited later,
                    because every deck referring to the old value would be
                    orphaned with nothing to notice.
                title:     { type: string, maxLength: 120 }
                sortOrder: { type: integer }
                active:    { type: boolean, default: true }
                accent:    { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
                secondary: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaxonomyTerm" }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: A term with this slug already exists for this kind
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /taxonomy/{kind}/unknown:
    parameters:
      - $ref: "#/components/parameters/TaxonomyKind"
    get:
      tags: [Taxonomy]
      summary: Values in use that no term defines
      description: |
        `decks.category`, `decks.industry` and `decks.source_type` are plain
        text with no foreign key, so a deck can carry a value no list contains
        — invisible to every filter that would otherwise find it. This is how
        those are found and reconciled deliberately, rather than by a migration
        that fails.
      responses:
        "200":
          description: Unknown values, most used first
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    value:     { type: string }
                    deckCount: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    InvalidInput:
      description: Malformed body, bad UUID, or a failed validation rule
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed, or expired token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - bearerAuth: []